requirements.txt
Django==2.0.2
djangorestframework==3.7.7
django-cors-headers==2.1.0
django-filter==1.1.0
factory-boy==2.10.0
imailing==1.0.0
coreapi==2.3.3
Pillow==5.2.0
